backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Pydantic for data validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# AI Provider SDKs
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
ollama>=0.1.6

# HTTP client for async operations
httpx>=0.26.0
aiohttp>=3.9.0

# Report generation
markdown>=3.5.2
weasyprint>=60.2  # For PDF generation

# Utilities
python-dotenv>=1.0.0
uuid6>=2024.1.12

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For TestClient

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
